requirements.txt
asgiref==3.8.1
Django==4.2.13
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==24.2
django-jazzmin==3.0.0
django-model-utils==4.5.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-excel==2.4.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
openpyxl==3.1.2
packaging==24.0
psycopg==3.1.18
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
